log files and success files management improvement

+++ scripts/sh/bamboo/validateInScratchOrg.sh
@@ -10,9 +10,18 @@
 
 mkdir -p log
 LOG_FILE=log/validateInScratchOrg.txt
-rm ${LOG_FILE}
+SUCCESS_FILE=successScratchOrg.tmp
+if [ -f $LOG_FILE ];
+then
+    rm ${LOG_FILE}
+fi
 exec > >(tee -a ${LOG_FILE} )
 exec 2> >(tee -a ${LOG_FILE} >&2)
+
+if [ -f SUCCESS_FILE ];
+then
+    rm ${SUCCESS_FILE}
+fi
 
 set -e
 
@@ -59,4 +68,4 @@ set +o
 echo "Validation Successfull!"
 
 # this is to let other scripts know that the build was successful
-touch successScratchOrg.tmp
+touch ${SUCCESS_FILE}
